v2.1/requirements.txt
Flask
flask-cors
google-generativeai
SpeechRecognition
pyaudio
werkzeug
python-docx
